requirements.txt
litellm
tenacity
datasets
sympy
antlr4-python3-runtime
word2number
latex2sympy2
timeout-decorator
jupyter
matplotlib
numpy
pandas
google-api-python-client
textblob
python-dotenv
requests
openai
anthropic
geopy
groq
loguru
pydantic
beautifulsoup4
transformers
torch
pytz
polygon-api-client
geopy
xmltodict
jaraco.windows
google-genai